refactor: move CLI agent resolution to runtime
Closes #5562
