+++ src/test-kernel/cli/AgentResolution.vitest.mts
@@ -43,8 +43,7 @@ describe('CLI agent resolution', () => {
   it('loads the local registry and returns a local agent for signed-out users', async () => {
     const local = agent('lean');
     mocks.getAgent.mockReturnValue(local);
-    const { resolveCliAgent } =
-      await import('@cli/commands/_helpers/agentResolution');
+    const { resolveCliAgent } = await import('@cli/runtime/agentResolution');
 
     await expect(resolveCliAgent('lean')).resolves.toBe(local);
 
@@ -56,8 +55,7 @@ describe('CLI agent resolution', () => {
   it('does a full registry load when the local registry misses', async () => {
     const remote = agent('orchestrator', 'remote');
     mocks.getAgent.mockReturnValueOnce(undefined).mockReturnValueOnce(remote);
-    const { resolveCliAgent } =
-      await import('@cli/commands/_helpers/agentResolution');
+    const { resolveCliAgent } = await import('@cli/runtime/agentResolution');
 
     await expect(resolveCliAgent('orchestrator')).resolves.toBe(remote);
 
@@ -73,8 +71,7 @@ describe('CLI agent resolution', () => {
     const remote = agent('lean', 'remote');
     mocks.authProvider.isAuthenticated.mockResolvedValue(true);
     mocks.getAgent.mockReturnValueOnce(local).mockReturnValueOnce(remote);
-    const { resolveCliAgent } =
-      await import('@cli/commands/_helpers/agentResolution');
+    const { resolveCliAgent } = await import('@cli/runtime/agentResolution');
 
     await expect(resolveCliAgent('lean')).resolves.toBe(remote);
 
@@ -89,8 +86,7 @@ describe('CLI agent resolution', () => {
     const local = agent('local:lean');
     mocks.authProvider.isAuthenticated.mockResolvedValue(true);
     mocks.getAgent.mockReturnValue(local);
-    const { resolveCliAgent } =
-      await import('@cli/commands/_helpers/agentResolution');
+    const { resolveCliAgent } = await import('@cli/runtime/agentResolution');
 
     await expect(resolveCliAgent('local:lean')).resolves.toBe(local);
