+++ src/test-kernel/cli/AgentsRunCommand.vitest.mts
@@ -42,7 +42,7 @@ vi.mock('@cli/commands/_helpers/output', () => ({
   emitCliResult: vi.fn(),
 }));
 
-vi.mock('@cli/commands/_helpers/agentResolution', () => ({
+vi.mock('@cli/runtime/agentResolution', () => ({
   resolveCliAgent: mocks.resolveCliAgent,
 }));
